fix(favorites) : #MAG-416 fix favorites not working and spacing with board title
