+++ frontend/src/providers/SVGProvider/types.ts
@@ -1,0 +1,9 @@
+import { ReactNode } from "react";
+
+export interface SVGProviderProps {
+  children: ReactNode;
+}
+
+export type SVGContextType = {
+  svgDoc: Document;
+};
