+++ frontend/src/routes/board/index.tsx
@@ -1,10 +1,13 @@
 import { BoardView } from "~/components/board-view/BoardView";
 import { BoardProvider } from "~/providers/BoardProvider";
+import { SVGProvider } from "~/providers/SVGProvider";
 
 export const App = () => {
   return (
     <BoardProvider>
-      <BoardView />
+      <SVGProvider>
+        <BoardView />
+      </SVGProvider>
     </BoardProvider>
   );
 };
